Fixed all remaining warnings.
git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop@2926 70805675-57b6-444b-bb00-18ecffcb4be8

+++ cru/guessio/guessio_cru.cpp
@@ -1237,7 +1237,7 @@ void CRUInput::getsowingdates(Gridcell& gridcell) {
 		for(i=0; i<npft; i++) {
 			if(pftlist[i].cftid >= 0 && pftlist[i].forcesowingdate)	{ //natural pft:s have cftid=-1	
 #if defined DYNAMIC_LANDCOVER_INPUT
-				gridcell.pft[i].sdate_force = sdates.Get(year,pftlist[i].name);
+				gridcell.pft[i].sdate_force = (int)sdates.Get(year,pftlist[i].name);
 #endif
 			}
 		}
@@ -1257,7 +1257,7 @@ void CRUInput::getharvestdates(Gridcell& gridcell) {
  		for(i=0; i<npft; i++)	{
 			if(pftlist[i].cftid >= 0 && pftlist[i].forceharvestdate) {	//natural pft:s have cftid=-1			
 #if defined DYNAMIC_LANDCOVER_INPUT
-				gridcell.pft[pftlist[i].id].hdate_force = hdates.Get(year,pftlist[i].name);
+				gridcell.pft[pftlist[i].id].hdate_force = (int)hdates.Get(year,pftlist[i].name);
 #endif
 			}
 		}
